requirements/base.txt
adrf==0.1.8
asgiref==3.8.1
async-property==0.2.2
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
gevent==24.11.1
geventhttpclient==2.3.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.32.2
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
psutil==6.1.0
psycopg2-binary==2.9.10
pyzmq==26.2.0
requests==2.32.3
setuptools==75.5.0
sqlparse==0.5.2
urllib3==2.2.3
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.1.1
Authlib==1.3.2
git+https://github.com/HuyNguyenDinh/django_sso_client.git